refatora cadastro cliente e adiciona persistencia de endereco e telefones

+++ src/main/java/com/oficinamecanica/oficina_mecanica_api/controller/GenericController.java
@@ -6,7 +6,7 @@ import java.net.URI;
 
 public interface GenericController {
 
-    default URI gerarHeaderLocationUri(int id) {
+    default URI gerarHeaderLocationUri(Long id) {
         return ServletUriComponentsBuilder
                 .fromCurrentRequestUri()
                 .path("/{id}")
